Initial batch of comments

+++ src/main/java/org/vexillum/CUSHION_SIZE.java
@@ -1,5 +1,8 @@
 package org.vexillum;
 
+/**
+ * An {@code enum} used to store information about cushion sizes and their cost
+ */
 public enum CUSHION_SIZE {
     SMALL(200),
     MEDIUM(1000),
@@ -12,7 +15,7 @@ public enum CUSHION_SIZE {
         this.value = value;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value / 100.0f;
     }
